tests/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile tests/requirements.in --output-file tests/requirements.txt
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
cachetools==5.5.2
    # via google-auth
certifi==2025.7.9
    # via
    #   kubernetes
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
cryptography==45.0.5
    # via -r tests/requirements.in
deepmerge==2.0
    # via -r tests/requirements.in
docker==7.1.0
    # via -r tests/requirements.in
durationpy==0.10
    # via kubernetes
execnet==2.1.1
    # via pytest-xdist
filelock==3.18.0
    # via -r tests/requirements.in
google-auth==2.40.3
    # via kubernetes
gprof2dot==2025.4.14
    # via pytest-profiling
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
jmespath==1.0.1
    # via -r tests/requirements.in
jsonschema==4.24.0
    # via -r tests/requirements.in
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via -r tests/requirements.in
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
packaging==25.0
    # via
    #   pytest
    #   pytest-rerunfailures
pluggy==1.6.0
    # via pytest
py==1.11.0
    # via pytest-forked
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==8.4.1
    # via
    #   -r tests/requirements.in
    #   pytest-forked
    #   pytest-pretty
    #   pytest-profiling
    #   pytest-rerunfailures
    #   pytest-testinfra
    #   pytest-xdist
pytest-forked==1.6.0
    # via -r tests/requirements.in
pytest-pretty==1.3.0
    # via -r tests/requirements.in
pytest-profiling==1.8.1
    # via -r tests/requirements.in
pytest-rerunfailures==15.1
    # via -r tests/requirements.in
pytest-testinfra==10.2.2
    # via -r tests/requirements.in
pytest-xdist==3.8.0
    # via -r tests/requirements.in
python-dateutil==2.9.0.post0
    # via kubernetes
pyyaml==6.0.2
    # via
    #   -r tests/requirements.in
    #   kubernetes
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   -r tests/requirements.in
    #   docker
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==14.0.0
    # via pytest-pretty
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   kubernetes
    #   pytest-profiling
    #   python-dateutil
typing-extensions==4.14.1
    # via referencing
urllib3==2.5.0
    # via
    #   docker
    #   kubernetes
    #   requests
websocket-client==1.8.0
    # via kubernetes
